refactor: unify useRegistrationData implementation
